Display events, add month label
Fixed MonthPane to display the events in the correct location
Changed MonthPane and WeekPane to VBoxes that have a GridPane. Added a
label to give the month.
TODO: Move day of week and month of year map from monthPane to model --
need to access this from weekPane also.
TODO: Add functionality to Edit button on BigEventPane
TODO: Make each day in the month a VBox so events don't stack on top of
each other.
TODO: Make the MonthPane GridPane expand to fill space.
TODO: Add scroll bar week view