Added printlns to show awkwardness on first day of month
event on the first ay of may is showing up on the right spot but in the
month April. The start date for that month gives three (april) and the
start date for the event gives 4 (may), but it's still being added to
april. I looked in the MonthPane addEvents and addEvent methods and
that's where the printlns are. Unsure if this has something to do with
the add/subtract months?